Fix ClassNotFoundException for SpecFailure class
For some reason I haven't taken the time to understand, one of the
implications of converting from Leiningen to tools.deps is that the
speclj.platform.Spec* classes don't get compiled. This results in a
ClassNotFoundException for speclj.platform.SpecFailure when I try to run
it.

I initially tried to just compile the classes with `javac` and commit
the class files to the repo. Although this felt like a hack, it worked
initially, but then when I tried to run speclj on a machine with a
different version of java installed, it complained about a version
mismatch in the class file.

Then I noticed that the classes were dead simple and could be ported to
a simple `deftype` definition in Clojure, eliminating the dependence on
a Java compilation step. But this didn't work because it wasn't a
Throwable so couldn't be thrown.

Then I tried `proxy`, but that didn't work because proxy creates
instances, not classes.

I realized that what I need is:

- a class
- whose instances are throwable (so extend Throwable or Exception)
- whose instances have a known type

I couldn't get (gen-class) or (ns ,,, (:gen-class ,,,)) to work until I
found this page: https://clojure.org/guides/deps_and_cli#aot_compilation

The thing I was missing is that I need a `classes/` directory to hold
the compiled classes, otherwise compilation will silently no-op.