Add macro like should= that also prints input form
Students didn't write the test suite that I'm using for the
Gradescope autograder, nor do they necessarily have access to it. It
would be nice for them to be able to see not just what their code
returned but also what was given to it so that they can more easily
figure out where a problem was.